requirements.txt
hidapi==0.14.0
PySimpleGUI==4.60.5
